Extract functions for ease of testing
